timer interrupt problem fix
When emulating large scale number of cores, we need to reduce timer frequency
to emualte timer event.  In linux, it uses PIT to calculate the frequency of
local apic timer's initial counter(register).  So in coremu, PIT_FREQ need to
be reduced to be (1193182 / coremu_get_thrs_per_core() * 2) instead of 1193182,
which is the real frequency PIT timer on i8254 chip.