Call mmio read function, disable cpu_exit, etc.
This commit contains too much stuff.

The mmio read is similar to the io port read problem. We need to call the actual
read function since it may change hardware status.

In replay, CPU knows when to jump out from TB link to handle interrupt, so no
need to call cpu_exit to unlink TB.

And it's possible for cpu_halted return true in cpu_exec when the CPU actually
needs an interrupt. (Is this the only case after we disabled cpu_exit?) So we
need to check this before returning halted in cpu_exec.